backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
